a couple refactors
