+++ src/containers/TopNav/TopNavButton/TopNavButton.scss
@@ -1,0 +1,5 @@
+@import 'src/styles/mixins';
+
+.TopNavButton {
+  @include top-nav-button;
+}
