+++ src/containers/TopNav/TopNavPopover/TopNavPopover.scss
@@ -1,22 +1,7 @@
+@import 'src/styles/mixins';
+
 .TopNavPopover {
-  align-items: center;
-  background: transparent;
-  border: none;
-  color: var(--color-sail-gray-500);
-  cursor: pointer;
-  display: flex;
-  font-weight: 700;
-  justify-content: center;
-  outline: none;
-  transition: color 0.1s ease;
-
-  &:hover {
-    color: var(--color-primary);
-  }
-
-  &--active {
-    color: var(--color-primary);
-  }
+  @include top-nav-button;
 
   &__chevron-icon {
     margin-left: 3px;
