+++ src/styles/_mixins.scss
@@ -4,3 +4,24 @@
   word-break: break-word;
   word-wrap: break-word;
 }
+
+@mixin top-nav-button {
+  align-items: center;
+  background: transparent;
+  border: none;
+  color: var(--color-sail-gray-500);
+  cursor: pointer;
+  display: flex;
+  font-weight: 700;
+  justify-content: center;
+  outline: none;
+  transition: color 0.1s ease;
+
+  &:hover {
+    color: var(--color-primary);
+  }
+
+  &--active {
+    color: var(--color-primary);
+  }
+}
